# primitives: Add initial proof-of-work funcs.