~**This requires #2787**.~

This introduces a primitives package that currently only contains functions related to proof of work along with comprehensive tests, benchmarks, full documentation and a skeleton `README.md`.

These functions are all the semantic equivalents of their respective functions in `blockchain/standalone` updated to use, accept, and return the new uint256 type instead of stdlib big integers.

Specifically, the following functions are updated and added along with their associated tests and benchmarks:

- `DiffBitsToUint256` (equiv to `standalone.CompactToBig`)
- `Uint256ToDiffBits` (equiv to `standalone.BigToCompact`)
- `CalcWork` (equiv to `standalone.CalcWork`)
- `HashToUint256`  (equiv to `standalone.HashToBig`)
- `CheckProofOfWorkRange`  (equiv to `standalone.CheckProofOfWorkRange`)
- `CheckProofOfWork`  (equiv to `standalone.CheckProofOfWork`)

It is split into multiple commits to ease the review process and each commit further describes the relevant changes.

A few key points:

The primary motivation for renaming the functions to specifically call out difficulty bits instead of the shorter "Compact" is to prevent future ambiguity with Satoshi-style variable length compact integers that will ultimately also be available from the package.

For legacy reasons, even though the difficulty bits encoding is only used in Decred to encode unsigned 256-bit integers which represent difficulty targets, the encoding is capable of representing negative numbers as well as numbers much larger than the maximum value of an unsigned 256-bit integer.  Therefore, in order to enable proper error detection and stay consistent with legacy code, `DiffBitsToUint256` returns additional flags to indicate the aforementioned conditions.

The original calculation that `CalcWork` performed (and therefore is used in consensus and must stay the same) involves a dividend of 2^256 which is not directly representable by a uint256, so this implementation retains the same semantics by transforming the calculation as described in detail by the comments in the code.

## Benchmarks

The following is a comparison between the existing implementations in the blockchain/standalone module with big integers (old) and the new implementation (new) averaging 10 runs each:

Name              |  Old Time/Op  |  New Time/Op  | Delta
------------------|---------------|---------------|--------
DiffBitsToUint256 | 158ns ± 1%    | 4ns ± 1%      | -97.35%
Uint256ToDiffBits | 151ns ± 2%    | 8ns ± 4%      | -94.82%
CalcWork          | 638ns ± 1%    | 59ns ± 1%     | -90.78%
HashToUint256     | 127ns ± 3%    | 3ns ± 1%      | -97.68%
CheckProofOfWork  | 314ns ± 1%    | 43ns ± 0%     | -86.42%

Name              | Old Allocs/Op | New Allocs/Op | Delta
------------------|---------------|---------------|---------
DiffBitsToUint256 |   3.00 ± 0%   |    0.00       | -100.00%
Uint256ToDiffBits |   2.00 ± 0%   | 0.00          | -100.00%
CalcWork          |   7.00 ± 0%   | 0.00          | -100.00%
HashToUint256     |   2.00 ± 0%   | 0.00          | -100.00%
CheckProofOfWork  |   5.00 ± 0%   | 0.00          | -100.00%

This is work towards #2786.