@davecgh 2021-11-13
> Commit title typo `rimitives: Add check proof of work. `.

Corrected.